undo device stats :(
