+++ src/lightning_lite/accelerators/cuda.py
@@ -11,14 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
-import shutil
-import subprocess
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import torch
 
-import lightning_lite.utilities.device_parser
 from lightning_lite.accelerators.accelerator import Accelerator
 from lightning_lite.utilities import device_parser
 
@@ -35,21 +31,6 @@ class CUDAAccelerator(Accelerator):
         if device.type != "cuda":
             raise ValueError(f"Device should be CUDA, got {device} instead.")
         torch.cuda.set_device(device)
-
-    def get_device_stats(self, device: torch.device) -> Dict[str, Any]:
-        """Gets stats for the given GPU device.
-
-        Args:
-            device: GPU device for which to get stats
-
-        Returns:
-            A dictionary mapping the metrics to their values.
-
-        Raises:
-            FileNotFoundError:
-                If nvidia-smi installation not found
-        """
-        return torch.cuda.memory_stats(device)
 
     def teardown(self) -> None:
         # clean up memory
@@ -81,61 +62,3 @@ class CUDAAccelerator(Accelerator):
             cls,
             description=f"{cls.__class__.__name__}",
         )
-
-
-def get_nvidia_gpu_stats(device: torch.device) -> Dict[str, float]:  # pragma: no-cover
-    """Get GPU stats including memory, fan speed, and temperature from nvidia-smi.
-
-    Args:
-        device: GPU device for which to get stats
-
-    Returns:
-        A dictionary mapping the metrics to their values.
-
-    Raises:
-        FileNotFoundError:
-            If nvidia-smi installation not found
-    """
-    nvidia_smi_path = shutil.which("nvidia-smi")
-    if nvidia_smi_path is None:
-        raise FileNotFoundError("nvidia-smi: command not found")
-
-    gpu_stat_metrics = [
-        ("utilization.gpu", "%"),
-        ("memory.used", "MB"),
-        ("memory.free", "MB"),
-        ("utilization.memory", "%"),
-        ("fan.speed", "%"),
-        ("temperature.gpu", "°C"),
-        ("temperature.memory", "°C"),
-    ]
-    gpu_stat_keys = [k for k, _ in gpu_stat_metrics]
-    gpu_query = ",".join(gpu_stat_keys)
-
-    index = torch._utils._get_device_index(device)
-    gpu_id = _get_gpu_id(index)
-    result = subprocess.run(
-        [nvidia_smi_path, f"--query-gpu={gpu_query}", "--format=csv,nounits,noheader", f"--id={gpu_id}"],
-        encoding="utf-8",
-        capture_output=True,
-        check=True,
-    )
-
-    def _to_float(x: str) -> float:
-        try:
-            return float(x)
-        except ValueError:
-            return 0.0
-
-    s = result.stdout.strip()
-    stats = [_to_float(x) for x in s.split(", ")]
-    gpu_stats = {f"{x} ({unit})": stat for (x, unit), stat in zip(gpu_stat_metrics, stats)}
-    return gpu_stats
-
-
-def _get_gpu_id(device_id: int) -> str:
-    """Get the unmasked real GPU IDs."""
-    # All devices if `CUDA_VISIBLE_DEVICES` unset
-    default = ",".join(str(i) for i in range(lightning_lite.utilities.device_parser.num_cuda_devices()))
-    cuda_visible_devices = os.getenv("CUDA_VISIBLE_DEVICES", default=default).split(",")
-    return cuda_visible_devices[device_id].strip()
