+++ src/lightning_lite/accelerators/tpu.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import torch
 
@@ -25,26 +25,6 @@ class TPUAccelerator(Accelerator):
 
     def init_device(self, device: torch.device) -> None:
         pass
-
-    def get_device_stats(self, device: torch.device) -> Dict[str, Any]:
-        """Gets stats for the given TPU device.
-
-        Args:
-            device: TPU device for which to get stats
-
-        Returns:
-            A dictionary mapping the metrics (free memory and peak memory) to their values.
-        """
-        import torch_xla.core.xla_model as xm
-
-        memory_info = xm.get_memory_info(device)
-        free_memory = memory_info["kb_free"]
-        peak_memory = memory_info["kb_total"] - free_memory
-        device_stats = {
-            "avg. free memory (MB)": free_memory,
-            "avg. peak memory (MB)": peak_memory,
-        }
-        return device_stats
 
     def teardown(self) -> None:
         pass
