+++ src/lightning_lite/utilities/imports.py
@@ -35,7 +35,6 @@ _HIVEMIND_AVAILABLE = package_available("hivemind")
 _HOROVOD_AVAILABLE = module_available("horovod.torch")
 _OMEGACONF_AVAILABLE = package_available("omegaconf")
 _POPTORCH_AVAILABLE = package_available("poptorch")
-_PSUTIL_AVAILABLE = package_available("psutil")
 _XLA_AVAILABLE: bool = package_available("torch_xla")
 
 # TODO(lite): import this from the fairscale files once they move to lite package
