+++ src/pytorch_lightning/accelerators/cuda.py
@@ -13,14 +13,14 @@
 # limitations under the License.
 import logging
 import os
+import shutil
+import subprocess
 from typing import Any, Dict, List, Optional, Union
 
 import torch
 
 import pytorch_lightning as pl
-from lightning_lite.accelerators.cuda import get_nvidia_gpu_stats as new_get_nvidia_gpu_stats
-from lightning_lite.utilities import device_parser, rank_zero_deprecation
-from lightning_lite.utilities.types import _DEVICE
+from lightning_lite.utilities import device_parser
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
@@ -54,7 +54,7 @@ class CUDAAccelerator(Accelerator):
         devices = os.getenv("CUDA_VISIBLE_DEVICES", all_gpu_ids)
         _log.info(f"LOCAL_RANK: {local_rank} - CUDA_VISIBLE_DEVICES: [{devices}]")
 
-    def get_device_stats(self, device: _DEVICE) -> Dict[str, Any]:
+    def get_device_stats(self, device: torch.device) -> Dict[str, Any]:
         """Gets stats for the given GPU device.
 
         Args:
@@ -101,9 +101,59 @@ class CUDAAccelerator(Accelerator):
         )
 
 
-def get_nvidia_gpu_stats(device: torch.device) -> Dict[str, float]:
-    rank_zero_deprecation(
-        "`pytorch_lightning.accelerators.cuda.get_nvidia_gpu_stats` has been deprecated in v1.8.0 and will be removed"
-        " in v1.10.0. Please use `lightning_lite.accelerators.cuda.get_nvidia_gpu_stats` instead."
+def get_nvidia_gpu_stats(device: torch.device) -> Dict[str, float]:  # pragma: no-cover
+    """Get GPU stats including memory, fan speed, and temperature from nvidia-smi.
+
+    Args:
+        device: GPU device for which to get stats
+
+    Returns:
+        A dictionary mapping the metrics to their values.
+
+    Raises:
+        FileNotFoundError:
+            If nvidia-smi installation not found
+    """
+    nvidia_smi_path = shutil.which("nvidia-smi")
+    if nvidia_smi_path is None:
+        raise FileNotFoundError("nvidia-smi: command not found")
+
+    gpu_stat_metrics = [
+        ("utilization.gpu", "%"),
+        ("memory.used", "MB"),
+        ("memory.free", "MB"),
+        ("utilization.memory", "%"),
+        ("fan.speed", "%"),
+        ("temperature.gpu", "°C"),
+        ("temperature.memory", "°C"),
+    ]
+    gpu_stat_keys = [k for k, _ in gpu_stat_metrics]
+    gpu_query = ",".join(gpu_stat_keys)
+
+    index = torch._utils._get_device_index(device)
+    gpu_id = _get_gpu_id(index)
+    result = subprocess.run(
+        [nvidia_smi_path, f"--query-gpu={gpu_query}", "--format=csv,nounits,noheader", f"--id={gpu_id}"],
+        encoding="utf-8",
+        capture_output=True,
+        check=True,
     )
-    return new_get_nvidia_gpu_stats(device)
+
+    def _to_float(x: str) -> float:
+        try:
+            return float(x)
+        except ValueError:
+            return 0.0
+
+    s = result.stdout.strip()
+    stats = [_to_float(x) for x in s.split(", ")]
+    gpu_stats = {f"{x} ({unit})": stat for (x, unit), stat in zip(gpu_stat_metrics, stats)}
+    return gpu_stats
+
+
+def _get_gpu_id(device_id: int) -> str:
+    """Get the unmasked real GPU IDs."""
+    # All devices if `CUDA_VISIBLE_DEVICES` unset
+    default = ",".join(str(i) for i in range(device_parser.num_cuda_devices()))
+    cuda_visible_devices = os.getenv("CUDA_VISIBLE_DEVICES", default=default).split(",")
+    return cuda_visible_devices[device_id].strip()
