+++ tests/tests_pytorch/accelerators/test_gpu.py
@@ -18,6 +18,7 @@ import torch
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators import CUDAAccelerator
+from pytorch_lightning.accelerators.cuda import get_nvidia_gpu_stats
 from pytorch_lightning.demos.boring_classes import BoringModel
 from tests_pytorch.helpers.runif import RunIf
 
@@ -27,6 +28,16 @@ def test_get_torch_gpu_stats(tmpdir):
     current_device = torch.device(f"cuda:{torch.cuda.current_device()}")
     gpu_stats = CUDAAccelerator().get_device_stats(current_device)
     fields = ["allocated_bytes.all.freed", "inactive_split.all.peak", "reserved_bytes.large_pool.peak"]
+
+    for f in fields:
+        assert any(f in h for h in gpu_stats.keys())
+
+
+@RunIf(min_cuda_gpus=1)
+def test_get_nvidia_gpu_stats(tmpdir):
+    current_device = torch.device(f"cuda:{torch.cuda.current_device()}")
+    gpu_stats = get_nvidia_gpu_stats(current_device)
+    fields = ["utilization.gpu", "memory.used", "memory.free", "utilization.memory"]
 
     for f in fields:
         assert any(f in h for h in gpu_stats.keys())
